grepc, grepc.1: Remove -tu from the defaults
Normally, one either wants to see a declaration/definition, or its
uses, but not both at the same time.  Also, there are usually many
more use sites than than definitions, and therefore definitions
are shadowed between all the noise.  Restrict the default to
declarations and definitions, which in my experience seems to be
the most common use case.

Signed-off-by: Alejandro Colomar <alx.manpages@gmail.com>