Don't find files recursively
grep only one file (or stdin).

The current implementation doesn't allow filterin standard input, which
I miss some times.  I'm removing this feature now, to be able to rewrite
most of the program in a way that allows filtering stdin, and then will
consider adding back a recursive mode if necessary.

Also, don't remove the two blanks at the start of the output, for
consistency when calling this program via xargs(1), which will become
more common now that it only handles single files.

Signed-off-by: Alejandro Colomar <alx@kernel.org>