bin/grepc: Optimize
Call pcre2grep(1) only once, which allows removing named pipes.  This is
an important optimization, and will also allow accepting several file
names in the command line (in a future commit).  The source code is also
significantly simplified.

It has a caveat: this single call to pcre2grep(1) will consume more
resources, and will crash on certain input files.  The workaround is to
restrict the search to just one or a few types of code.  Document this
in the manual page.  This caveat only applies to very specific files,
which so far I've only found in the Linux kernel source tree.