bin/grepc, grepc.1: Always enable recursive mode; remove -r
The non-recursive behavior is so slow that it's quite useless.  It is
only useful when reading from standard input, but for several files it's
useless.  Make '-r' the default when any files are specified, and remove
the flag (since now it's always enabled).  This happens to bring back a
behavior close to what we had before
2fb22befdfb9 ("Don't find files recursively").

Non-seekable files can still be read, if redirected to standard input.

Update the example from the manual page to reflect current usage.

Signed-off-by: Alejandro Colomar <alx@kernel.org>