man/man2/mremap.2: Describe multiple-mapping move
Document the new behaviour introduced in Linux 6.17 whereby it is now
possible to move multiple mappings in a single operation, as long as the
operation is purely a move, that is, old_size == new_size and
MREMAP_FIXED is specified.

This change also explains the limitations of this method and the
possibility of partial failure.

Finally, we pluralise language where it makes sense to so the documentation
does not contradict either this new capability nor the pre-existing edge
case.

Example code is enclosed below demonstrating the behaviour which is now
possible:

	#define _GNU_SOURCE
	#include <err.h>
	#include <stddef.h>
	#include <stdlib.h>
	#include <sys/mman.h>
	#include <unistd.h>

	int
	main(void)
	{
		void    *ptr, *tgt_ptr;
		size_t  page_size;

		page_size = sysconf(_SC_PAGESIZE);
		ptr = mmap(NULL, 10 * page_size, PROT_READ | PROT_WRITE,
			       MAP_ANON | MAP_PRIVATE, -1, 0);
		if (ptr == MAP_FAILED)
			err(EXIT_FAILURE, "mmap");

		tgt_ptr = mmap(NULL, 10 * page_size, PROT_NONE,
			       MAP_ANON | MAP_PRIVATE, -1, 0);
		if (tgt_ptr == MAP_FAILED)
			err(EXIT_FAILURE, "mmap");

		/* Unmap every other page. */
		for (int i = 1; i < 10; i += 2)
			munmap(ptr + i * page_size, page_size);

		/* Now move all 5 distinct mappings to tgt_ptr. */
		ptr = mremap(ptr, 10 * page_size, 10 * page_size,
			     MREMAP_MAYMOVE | MREMAP_FIXED, tgt_ptr);
		if (ptr == MAP_FAILED)
			err(EXIT_FAILURE, "mremap");

		exit(EXIT_SUCCESS);
	}

Signed-off-by: Lorenzo Stoakes <lorenzo.stoakes@oracle.com>
Message-ID: <4e0c992a6374e417367475e3b3bbbc9d43380f4c.1754924278.git.lorenzo.stoakes@oracle.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>