man/man2/syscalls.2: Remove pread and pwrite from list of arch-specific syscalls
Current version says that pread(2) and pwrite(2) only ever existed in
avr32 and blackfin archs, which were removed from kernel.
This is not true.  pread(2) and pwrite(2) are present in modern Linux
for all archs.  They were merely renamed to pread64(2) and pwrite64(2)
in Linux 2.6, as explained in big table in this manual page.

Signed-off-by: Askar Safin <safinaskar@zohomail.com>
Message-ID: <20250819161000.768159-3-safinaskar@zohomail.com>
Signed-off-by: Alejandro Colomar <alx@kernel.org>