man/: Address diagnostic about mismatched quotes
Most of these are false positives, in the sense that a line break is
escaped, and the matching quote is in the next source line.

However, let's remove those escaped line breaks, which make reading the
source more difficult.  For this, rename some parameters to be shorter,
and allow some lines to go slightly past the 80-column right margin.

Signed-off-by: Alejandro Colomar <alx@kernel.org>