man/man2/open_tree{,_attr}.2: Document open_tree_attr(), and add link page
This is a new API added in Linux 6.15, and is effectively just a minor
expansion of open_tree(2) in order to allow for MOUNT_ATTR_IDMAP to be
changed for an existing ID-mapped mount.  glibc does not yet have a
wrapper for this.

While working on this man-page, I discovered a bug in open_tree_attr(2)
that accidentally permitted changing MOUNT_ATTR_IDMAP for extant
detached ID-mapped mount objects.  This is definitely a bug, but there
is no need to add this to BUGS because the patch to fix this has already
been accepted (slated for 6.18, and will be backported to 6.15+).

Cc: Christian Brauner <brauner@kernel.org>
Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>
Message-ID: <20250925-new-mount-api-v5-7-028fb88023f2@cyphar.com>
Reviewed-by: Askar Safin <safinaskar@gmail.com>
[alx: amend some examples: s/open_tree/&_attr/]
Signed-off-by: Alejandro Colomar <alx@kernel.org>