man/man3/[umax]abs.3: Document umaxabs(3)
Rename uimaxabs(3) => umaxabs(3).  And add a link page.

Keep a link with the old name, so that programmers looking for the old
name are redirected to the new name, so they find what they need.

I've removed documentation for uimaxabs(3), instead of keeping
information in HISTORY, as it has only existed for one glibc version,
to avoid programmers being confused by the old name.

Link: <https://www.open-std.org/jtc1/sc22/wg14/www/docs/n3577.txt>
Signed-off-by: Alejandro Colomar <alx@kernel.org>