src/bin/grepc: Use the same error code as grep(1)
Exit with 2 on error.

If nothing matches, it's a success, and it exits with 0 normally.
This is different from grep(1), but I don't know how to write
the script to exit with 1 if nothing matches.

This implementation may also ignore some xargs(1) errors.  A future
commit --with code suggested by Paul Eggert-- will fix this.

Link: <https://software.codidact.com/posts/294883>
Signed-off-by: Alejandro Colomar <alx@kernel.org>