src/bin/grepc_c: -tt, -tut: Reuse grepc_c_body_()
In the last case, it removes part of the regex, but I believe that
wasn't necessary.

Also, while at this, make the trailing of the regexes to be consistently
'[^;];', which is simpler, and seems to work just fine.  If we ever see
some false positives or negatives with this regex, we should fix them
all at once, and maybe put them in a helper function.

Signed-off-by: Alejandro Colomar <alx@kernel.org>