src/bin/grepc_c: Split grepc_c_body_() and grepc_c_u_body_()
grepc_c_u_body_() consumes a lot of resources unnecessarily.
This is especially true within grepc_c_t_td_braced(), which was crashing
with relatively small code.  The reason is that grepc_c_t_td_braced()
searches for code where the identifier is last, so it needs to store
a lot of code just in case.  For that, an optimized grepc_c_body_() is
more appropriate.

Signed-off-by: Alejandro Colomar <alx@kernel.org>