*/: Restore some old copyright notices
This commit partially reverts
9f2986c34166 (2025-05-06; "*, CREDITS: Unify copyright notices").

Some authors have given express permission for the transformation done
in that old commit, and their notices remain transformed.  But those
that didn't give express permission have their copyright notices
restored now.

It's unclear whether the transformation was legally valid.  I believe it
was, but since IANAL, and just to be cautious, we've agreed to restore
them.

This restoration, at least, unifies the formatting, to reduce entropy.

Link: <https://lore.kernel.org/linux-man/u2ogua4573d2xm2p2oiuna67kydkr3e26pt6lixeidezdw34dg@nvn64na3cptt/>
Signed-off-by: Alejandro Colomar <alx@kernel.org>