man/man2/semop.2: VERSIONS: Fix incorrect history
Split semop(2) and semtimedop(2) in this section, and move the
previously listed standards/versions to the corresponding functions.

The VERSIONS section is really a HISTORY section that seems to have
mixed up semop(2) and semtimedop(2).  The relevant commit (4131356cdab8)
removed some context specifying which standards/versions apply to
semtimedop(2) only, causing this mix-up.  The diff for this commit on
man/man2/semop.2 is given below.

    $ git diff -U1 4131356cdab8^! man2/semop.2
    diff --git a/man2/semop.2 b/man2/semop.2
    index 7a1416a26..fe2906be4 100644
    --- a/man2/semop.2
    +++ b/man2/semop.2
    @@ -357,11 +357,8 @@ .SH ERRORS
     .IR semval .
    -.SH VERSIONS
    -.BR semtimedop ()
    -first appeared in Linux 2.5.52,
    -and was subsequently backported into Linux 2.4.22.
    -glibc support for
    -.BR semtimedop ()
    -first appeared in Linux 2.3.3.
     .SH STANDARDS
    -POSIX.1-2001, POSIX.1
    -2008, SVr4.
    +POSIX.1-2008.
    +.SH VERSIONS
    +Linux 2.5.52 (backported into Linux 2.4.22),
    +glibc 2.3.3.
    +POSIX.1-2001, SVr4.
     .\" SVr4 documents additional error conditions EINVAL, EFBIG, ENOSPC.

The removed information is restored on account of the semop(2) and
semtimedop(2) functions being described separately, removing the mix-up.

Fixes: 4131356cdab8 (2023-03-17; "man*/, man-pages.7: VERSIONS, STANDARDS, HISTORY: Reorganize sections")
Signed-off-by: Seth McDonald <sethmcmail@pm.me>
Message-ID: <86b4c2ec94334aadcc8abc9f17b34b60716d399b.1766641593.git.sethmcmail@pm.me>
[alx: split commit in two]
Signed-off-by: Alejandro Colomar <alx@kernel.org>